feat(copy-paste-btn-directive): refactor, split directives
