+++ packages/components/copy-paste-directive/dashboard-copy-clipboard-dialog/dashboard-copy-clipboard-dialog.component.scss
