+++ packages/components/copy-paste-directive/index.ts
@@ -1,4 +1,5 @@
-export { CopyPasteBtnDirective } from './copy-paste-btn.directive';
+export { CopyDirective } from './copy.directive';
+export { PasteDirective } from './paste.directive';
 export { DashboardCopyClipboardDialogComponent } from './dashboard-copy-clipboard-dialog/dashboard-copy-clipboard-dialog.component';
 export {
     DEFAULT_CONFIG,
