+++ packages/components/copy-paste-directive/ng-package.json
